requirements.txt
blinker==1.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
greenlet==3.0.3
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.0
pandas==2.2.2
patsy==0.5.6
plotly==5.21.0
plotly-express==0.4.1
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
scipy==1.13.0
six==1.16.0
SQLAlchemy==2.0.29
statsmodels==0.14.2
tenacity==8.2.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.2
